rename ui-files in client_app"
